Pulled connectionStatus into operations folder
