+++ src/client/src/configureStore.ts
@@ -4,10 +4,10 @@ import { combineEpics, createEpicMiddleware } from 'redux-observable'
 
 import { ApplicationDependencies } from './applicationServices'
 import rootReducer, { GlobalState } from './combineReducers'
-import { connectionStatusEpic } from './connectionStatusOperations'
 import { linkEpic } from './linkEpic'
 import { openfinEpic } from './openfinEpics'
 import { compositeStatusServiceEpic } from './operations/compositeStatusService'
+import { connectionStatusEpic } from './operations/connectionStatus'
 import { pricingServiceEpic } from './operations/pricing'
 import { referenceServiceEpic } from './referenceDataOperations'
 import { analyticsServiceEpic } from './ui/analytics'
