+++ src/client/src/main.tsx
@@ -5,7 +5,7 @@ import { timer } from 'rxjs'
 import { createApplicationServices } from './applicationServices'
 import { getEnvVars } from './config/config'
 import configureStore from './configureStore'
-import { ConnectionActions } from './connectionActions'
+import { ConnectionActions } from './operations/connectionStatus'
 import { FakeUserRepository } from './services'
 import { OpenFin } from './services/openFin'
 import { AutobahnConnectionProxy, logger } from './system'
