+++ src/client/src/openfinEpics.ts
@@ -2,7 +2,7 @@ import { combineEpics, ofType } from 'redux-observable'
 import { bindCallback } from 'rxjs'
 import { filter, map, mergeMap, switchMapTo, takeUntil, tap, withLatestFrom } from 'rxjs/operators'
 import { ApplicationEpic } from './ApplicationEpic'
-import { ACTION_TYPES as CONNECTION_ACTION_TYPES } from './connectionActions'
+import { ACTION_TYPES as CONNECTION_ACTION_TYPES } from './operations/connectionStatus'
 import { openfinServiceEpics } from './services/openFin'
 import { Direction, SpotPriceTick } from './types'
 import { CurrencyPair } from './types/currencyPair'
