+++ src/client/src/ui/blotter/reducer.ts
@@ -1,5 +1,5 @@
 import * as keyBy from 'lodash.keyby'
-import { ACTION_TYPES as CONNECTION_ACTION_TYPES, DisconnectAction } from '../../connectionActions'
+import { ACTION_TYPES as CONNECTION_ACTION_TYPES, DisconnectAction } from '../../operations/connectionStatus'
 import { Trade } from '../../types'
 import { ACTION_TYPES, BlotterActions } from './actions'
